Fix display function by removing broken cache
The attempted approach to caching results was broken because display is
expected to return a string, not a meta-result. Thus, when display was
applied to a string, it created but did not return a new meta-result
table, only returning nil strings.

This could be fixed pretty easily by changing the use of display to
expect a meta-result, but I've opted to leave that as a possible future
optimization.